Fix sort order of movies in volume list

+++ lib/models/firestore_great_movie_model.dart
@@ -14,7 +14,7 @@ class FirestoreGreatMovie {
   bool? isCriterion;
   bool? isWatched;
   String? dateWatched;
-  int? userStarRating;
+  double? userStarRating;
   String? userReview;
 
   FirestoreGreatMovie(
